forecasto-agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "forecasto-agent"
version = "0.1.0"
description = "Forecasto Agent — monitors folders and processes documents with AI"
requires-python = ">=3.11"
dependencies = [
    "watchdog>=4.0",
    "anthropic>=0.40",
    "httpx>=0.27",
    "pymupdf>=1.24",          # PDF → images
    "Pillow>=10.0",
    "pystray>=0.19",          # System tray (Linux/Mac/Windows)
    "plyer>=2.1",             # Desktop notifications
    "click>=8.1",             # CLI
    "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
dev = ["ruff", "mypy"]

[project.scripts]
forecasto-agent = "forecasto_agent.main:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/forecasto_agent"]
